FISH-6160 Remove CatalinaLogger and IASLogger and bundle into LoggerBase - they're barely different
